+++ okx/Convert.py
@@ -10,16 +10,22 @@ class ConvertAPI(OkxClient):
         params = {}
         return self._request_with_params(GET, GET_CURRENCIES, params)
 
-    def get_currency_pair(self, fromCcy='', toCcy=''):
+    def get_currency_pair(self, fromCcy='', toCcy='', convertMode=None):
         params = {"fromCcy": fromCcy, 'toCcy': toCcy}
+        if convertMode is not None:
+            params['convertMode'] = convertMode
         return self._request_with_params(GET, GET_CURRENCY_PAIR, params)
 
-    def estimate_quote(self, baseCcy = '', quoteCcy = '', side = '', rfqSz = '', rfqSzCcy = '', clQReqId = '',tag=''):
+    def estimate_quote(self, baseCcy = '', quoteCcy = '', side = '', rfqSz = '', rfqSzCcy = '', clQReqId = '',tag='', convertMode=None):
         params = {'baseCcy': baseCcy, 'quoteCcy': quoteCcy, 'side':side, 'rfqSz':rfqSz, 'rfqSzCcy':rfqSzCcy, 'clQReqId':clQReqId,'tag':tag}
+        if convertMode is not None:
+            params['convertMode'] = convertMode
         return self._request_with_params(POST, ESTIMATE_QUOTE, params)
 
-    def convert_trade(self, quoteId = '', baseCcy = '', quoteCcy = '', side = '', sz = '', szCcy = '', clTReqId = '',tag=''):
+    def convert_trade(self, quoteId = '', baseCcy = '', quoteCcy = '', side = '', sz = '', szCcy = '', clTReqId = '',tag='', convertMode=None):
         params = {'quoteId': quoteId, 'baseCcy': baseCcy, 'quoteCcy':quoteCcy, 'side':side, 'sz':sz, 'szCcy':szCcy, 'clTReqId':clTReqId,'tag':tag}
+        if convertMode is not None:
+            params['convertMode'] = convertMode
         return self._request_with_params(POST, CONVERT_TRADE, params)
 
     def get_convert_history(self, after = '', before = '', limit = '',tag=''):